+++ tapis-systemsapi/src/main/java/edu/utexas/tacc/tapis/systems/api/requests/ReqPostSystem.java
@@ -50,12 +50,12 @@ public final class ReqPostSystem
   public String dtnMountSourcePath;
   public boolean isDtn;
   public boolean canExec;
+  public boolean canRunBatch;
   public List<JobRuntime> jobRuntimes;
   public String jobWorkingDir;
   public List<KeyValuePair> jobEnvVariables;
   public int jobMaxJobs = DEFAULT_JOBMAXJOBS;
   public int jobMaxJobsPerUser = DEFAULT_JOBMAXJOBSPERUSER;
-  public boolean jobIsBatch;
   public SchedulerType batchScheduler;
   public List<LogicalQueue> batchLogicalQueues;
   public String batchDefaultLogicalQueue;