+++ tapis-systemslib/src/main/java/edu/utexas/tacc/tapis/systems/model/PatchSystem.java
@@ -27,12 +27,12 @@ public final class PatchSystem
   private final String dtnSystemId;
   private final String dtnMountPoint;
   private final String dtnMountSourcePath;
+  private final Boolean canRunBatch;
   private final List<JobRuntime> jobRuntimes;
   private final String jobWorkingDir;
   private final List<KeyValuePair> jobEnvVariables;
   private final Integer jobMaxJobs;
   private final Integer jobMaxJobsPerUser;
-  private final Boolean jobIsBatch;
   private final SchedulerType batchScheduler;
   private final List<LogicalQueue> batchLogicalQueues;
   private final String batchDefaultLogicalQueue;
@@ -54,7 +54,7 @@ public final class PatchSystem
                      Integer port1, Boolean useProxy1, String proxyHost1, Integer proxyPort1,
                      String dtnSystemId1, String dtnMountPoint1, String dtnMountSourcePath1,
                      List<JobRuntime> jobRuntimes1, String jobWorkingDir1, List<KeyValuePair> jobEnvVariables1,
-                     Integer jobMaxJobs1, Integer jobMaxJobsPerUser1, Boolean jobIsBatch1,
+                     Integer jobMaxJobs1, Integer jobMaxJobsPerUser1, Boolean canRunBatch1,
                      SchedulerType batchScheduler1, List<LogicalQueue> batchLogicalQueues1,
                      String batchDefaultLogicalQueue1, String batchSchedulerProfile1,
                      List<Capability> jobCapabilities1, String[] tags1, Object notes1, String importRefId1)
@@ -70,12 +70,12 @@ public final class PatchSystem
     dtnSystemId = dtnSystemId1;
     dtnMountPoint = dtnMountPoint1;
     dtnMountSourcePath = dtnMountSourcePath1;
+    canRunBatch = canRunBatch1;
     jobRuntimes = (jobRuntimes1 == null) ? null : new ArrayList<>(jobRuntimes1);
     jobWorkingDir = jobWorkingDir1;
     jobEnvVariables = (jobEnvVariables1 == null) ? null : new ArrayList<>(jobEnvVariables1);
     jobMaxJobs = jobMaxJobs1;
     jobMaxJobsPerUser = jobMaxJobsPerUser1;
-    jobIsBatch = jobIsBatch1;
     batchScheduler = batchScheduler1;
     batchLogicalQueues = (batchLogicalQueues1 == null) ? null : new ArrayList<>(batchLogicalQueues1);
     batchDefaultLogicalQueue = batchDefaultLogicalQueue1;
@@ -125,7 +125,7 @@ public final class PatchSystem
 
   public Integer getJobMaxJobsPerUser() { return jobMaxJobsPerUser; }
 
-  public Boolean getJobIsBatch() { return jobIsBatch; }
+  public Boolean getCanRunBatch() { return canRunBatch; }
 
   public SchedulerType getBatchScheduler() { return batchScheduler; }